withProvider wraps pages not after
when I was wrapping After with withProvider, I wasn't able to
access `req` and `res`, which I _need_ to get this thing to work
without JS.

Wrapping the pages means that withProvider always gets the `req`,
`res`, and `match` params on first load. Means I can check for
cookies and set them and all that.

//

Only pages have access to the `static async getInitialProps` method.
But since it was a flat hierarchy, it's not obvious which ones are
pages.

As such, I just created a folder for them.