switchTheme → setStore
switchTheme would reset one specific variable in the context for
us whereas we need something more general purpose.

setStore will update the context.store variable that gets passed
through as `context`, as well as it sets a `store` cookie in the
browser with the current updated this.state.context.store.

This means that on page reloads, we can update the
context.store using our store cookies.

This is going really well.