pull theme out of the store
since theme was always a derived value (ie, it was dependent
on the value of store.form.selectedTheme), then there's no
reason to keep it in the context.

this is a huge bonus for me, as it knocks out so much weird logic
in one fell swoop.

so many TODOs banished from the codebase.