validate values on server-side before saving them
Before, what was happening was that invalid values were being
saved when there was no JS on the page.

Since the validation method was in the page component itself,
we weren't validating any values before saving them -- instead,
we were saving anything that was a query parameter.

Once the page component had validated the values it was getting
back, it would render an error message to the page, but the values
had already been saved by that point.

The problem is that the `withProvider` component gets the `res` and
`req` parameters in its `getInitialProps` method but isn't able to
pass it down to the page components itself.

Therefore, the validation has to be pulled out of the page components
and done in the withProvider component itself.

Not a fan of the global JS files but this was the most workable
solution I could come up with.

//

There are genuine improvements here that are notable though:
- now only whitelisted query fields can be saved
  - the earlier behaviour was to save all query fields all the time
- now only validated values are saved
  - the earlier behaviour was to save values that would fail validation
- now invalidated values are reset
  - the earlier behaviour was to never be able to reset a value
    once it passed validation