keep validation in page components
in commit #4765729b8b6244b6d75dc467b18e825d21dc4fa7, I noted that
there wasn't a good way to pass the `req` or `res` objects to the
page component itself, _but_ the withProvider component is getting
the page components themselves because it is wrapping them.

Since they are also all being wrapped with `withContext`, I can't
directly access methods on the WrappedComponent (ie, I can't check
for `WrappedComponent.fields`), but we can still pass them through
to `withProvider` as optional parameters.

This means that we can create (static) validation functions inside
of our page components and then pass them up to withProvider, which
allows us to get rid of the `validation.js` file -- 🥂🙌

Now the components themselves are responsible for their validation
methods and query field whitelisting.

If either a list of fields or a validation function is missing,
then no query parameters will be saved as cookies.
Before, we were requiring fields to be whitelisted, but a validation
method was optional. So now we're being more deliberate about
the kinds of things we want to save, as well as disallowing saving
query parameters entirely on most pages.