Refactor getInitialProps method to prefer global field keys
If we get a query like ?language=en&field=value, we will just
throw away the non-global key and set the global cookie.

Added a bunch of tests for the new `returnKeyAndValue` function.

Also, I've moved all of the validation logic out of `setSSRCookie`
because it's handled by `WithProvider.validateCookie`.