Global language state is GO
Added a global language state switcher.

Unlike all of the other things we save to the store, this one
is a global value that's not associated with a specific page.

To underscore that, it's in the navigation so it can also be set
from any page.

Our previous strategy of validating fields on a page-by-page basis
falls apart with the global language value.

My solution to this puzzle was to create a global whitelist in
`WithProvider`, and then pull together all the validation logic
so that I could check for (i) global values, and then (ii) page-
specific values the same way every time -- whether through
client-side JS or through query parameters.

The LanguageSwitcher is the implementation of this idea.
It works on any page, and sets a global value whether JS is
enabled or not.

My implementation is still naive enough that if you created
a `/language` page in your app, then you need a different
global key, but this isn't a problem for our small web service.