Add logic to accept default language header from browser
We are checking cookies to see if someone has a default language
set.

However, if they come to our site for the first time:
- they won't have cookies
- they may have a language setting in their browser

So that's when we need to check.

If they do have a cookie set, then we don't check.
That way, we won't override an explicit decision. Easy peasy.