Port in code from rescheduler
I'm trying to keep this as in-sync as I can.

In the rescheduler, we added a couple extra comments, as well as
we're trimming values before validating and saving them.

So if someone enters a value such as:
- 'paul@pcraig3.ca   '
as their email address, we will trim it to 'paul@pcraig3.ca' and
then that's what we'll save.

Added an extra test for that case as well.