Remove cookie encryption
For our application, we don't need it.

- Cookies are only served to our rescheduler domain (means that other
  sites can't look at them).
- Our rescheduler domain only sends cookies over HTTPS, so the site
  traffic is encrypted end-to-end

There shouldn't be a way to actually get the content of the cookies
unless you're the person who entered the values in the first place.

If cookie values are being used to authenticate users or something,
then, yeah, they should be encrypted.

But ours aren't.