Never return empty key from pathname
On the landing page, the pathname is "/".
Since it would be trimmed in the `returnKeyAndValue` function,
we would be passing key === '' to `validateCookie`.
This meant that `validateCookie` would fail out.

In practical terms, this meant that visiting the root URL "/"
and passing in a query parameter would cause the language set in
the cookie to be ignored.