Only validate query parameters when a query param key matches a field on the page
If we added a query parameter to a URL and did a full-page reload,
you could lose previously entered data.

Now, only run a validation if at least one query key is in
the permitted fields.

Note that we are possibly losing error messaging if the `val`
passed in is a bad value, but we weren't using this in our
application for anything.

What we get instead is, arguably, more readable code.