Use POST not GET for sending values back in no-JS mode
It looks like it's easier than I thought it would be.

POST params are better in a way because the values are hidden
instead of being exposed in the URL string.
It also simplified the logic because now we are no longer checking
URL parameters when we reload our basic HTML page.