Look for query parameters if there are no POST parameters
The values we submit through the form we want to consume through
POST parameters, but the language is fine to be a query parameter still.

So this way, we will still keep the existing behaviour to do with
query parameters. But we won't show error messages if someone manually
'hacks' the URL to type in values.

Not a big D.